Add test helpers
