+++ tests/frontier/blockchain_st_test_helpers.py
@@ -207,24 +207,6 @@ def json_to_state(raw: Any) -> State:
     return state
 
 
-def load_test_transaction(
-    test_dir: str, test_file: str, network: str
-) -> Dict[str, Any]:
-    pure_test_file = os.path.basename(test_file)
-    test_name = os.path.splitext(pure_test_file)[0]
-    path = os.path.join(test_dir, test_file)
-
-    with open(path, "r") as fp:
-        json_data = json.load(fp)[f"{test_name}"]
-
-    tx_rlp = hex_to_bytes(json_data["txbytes"])
-    tx = rlp.decode_to(Transaction, tx_rlp)
-
-    test_result = json_data["result"][network]
-
-    return {"transaction": tx, "test_result": test_result}
-
-
 run_frontier_blockchain_st_tests = partial(
     run_blockchain_st_test, network="Frontier"
 )
